feat(kaze): Complete functionality of falconguard middleware and remove /register route
