+++ services/kaze/controllers/user.go
@@ -10,49 +10,6 @@ import (
 	"github.com/sdslabs/gasper/types"
 )
 
-// Register handles registration of new users
-func Register(c *gin.Context) {
-	user := &types.User{}
-	if err := c.BindJSON(user); err != nil {
-		c.AbortWithStatusJSON(400, gin.H{
-			"success": false,
-			"error":   err.Error(),
-		})
-		return
-	}
-
-	filter := types.M{mongo.EmailKey: user.Email}
-	count, err := mongo.CountUsers(filter)
-	if err != nil {
-		utils.SendServerErrorResponse(c, err)
-		return
-	}
-	if count > 0 {
-		c.JSON(400, gin.H{
-			"success": false,
-			"error":   "email already registered",
-		})
-		return
-	}
-
-	hashedPass, err := utils.HashPassword(user.GetPassword())
-	if err != nil {
-		utils.SendServerErrorResponse(c, err)
-		return
-	}
-	user.SetPassword(hashedPass)
-	user.SetAdmin(false)
-
-	if _, err = mongo.RegisterUser(user); err != nil {
-		utils.SendServerErrorResponse(c, err)
-		return
-	}
-	c.JSON(200, gin.H{
-		"success": true,
-		"message": "user created",
-	})
-}
-
 // GetUserInfo gets info regarding particular user
 func GetUserInfo(c *gin.Context) {
 	user, err := mongo.FetchSingleUserWithoutPassword(c.Param("user"))
